backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
httpx
google-generativeai
pytest
python-dotenv
boto3
google-auth
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-huggingface
faiss-cpu
sentence-transformers
PyPDF
lark
unstructured
# Additional packages for AI Agent enhancement
langchain-experimental
openai
anthropic
tiktoken
numpy
pandas
matplotlib
seaborn
plotly
markdown
# markdown-pdf  # Temporarily disabled due to compatibility issues